LPS-112056 Move ResourceActionsBag to upper level and public package
